+++ packages/stream_chat_flutter/CHANGELOG.md
@@ -8,7 +8,7 @@
 
 🔄 Changed
 
-- [[#1125]](https://github.com/GetStream/stream-chat-flutter/issues/1125) `defaultUserImage`, `placeholderUserImage`, and `reactionIcons` have been refactored out of `StreamChatThemeData` and into the new `StreamChatConfiguration` class.
+- [[#1125]](https://github.com/GetStream/stream-chat-flutter/issues/1125) `defaultUserImage`, `placeholderUserImage`, and `reactionIcons` have been refactored out of `StreamChatThemeData` and into the new `StreamChatConfigurationData` class.
 
 ## 5.0.0-beta.1